psa: move psa to its own package for reuse
