+++ pkg/virt-controller/watch/BUILD.bazel
@@ -8,7 +8,6 @@ go_library(
         "migrationpolicy.go",
         "node.go",
         "pool.go",
-        "psa.go",
         "replicaset.go",
         "util.go",
         "vm.go",
@@ -32,6 +31,7 @@ go_library(
         "//pkg/monitoring/vmstats:go_default_library",
         "//pkg/network/sriov:go_default_library",
         "//pkg/network/vmispec:go_default_library",
+        "//pkg/psa:go_default_library",
         "//pkg/service:go_default_library",
         "//pkg/storage/export/export:go_default_library",
         "//pkg/storage/snapshot:go_default_library",
@@ -109,7 +109,6 @@ go_test(
         "migration_test.go",
         "node_test.go",
         "pool_test.go",
-        "psa_test.go",
         "replicaset_test.go",
         "vm_test.go",
         "vmi_test.go",
